requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --no-emit-index-url requirements/requirements.in
#
aiohttp==3.9.3
    # via discord-py
aiosignal==1.3.1
    # via aiohttp
attrs==23.2.0
    # via aiohttp
discord-py==2.3.2
    # via -r requirements/requirements.in
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
idna==3.6
    # via yarl
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
secretbox==2.8.0
    # via -r requirements/requirements.in
yarl==1.9.4
    # via aiohttp
